Architecture Refactoring and sass style extension introduction
