--- src/App.css
@@ -1,21 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.button {
-  outline: none !important;
-}
-
-.button:hover,
-.button:active {
-    background-color: #d2f6f9 !important;
-    color: #007da0 !important;
-}
-
-.red {
-  color: red;
-}
-
-.bold {
-  font-weight: bold;
-}
