+++ src/containers/App.js
@@ -1,0 +1,66 @@
+import React, { Component } from 'react';
+import classes from './App.css';
+import Persons from '../components/Persons/Persons';
+import Cockpit from '../components/Cockpit/Cockpit';
+
+class App extends Component {
+  state = {
+    persons: [
+      {id: 'asdf1', name: 'Francesco', age: 33},
+      {id: 'asdf2', name: 'Gabriele', age: 33},
+      {id: 'asdf3', name: 'Chiara', age: 27}
+    ],
+    otherState: 'some other value',
+    showPersons: false
+  }
+
+  nameChangedHandler = (event, id) => {
+    const personIndex = this.state.persons.findIndex(p => {
+      return p.id === id;
+    });
+
+    const person = {
+      ...this.state.persons[personIndex]
+    };
+
+    person.name = event.target.value;
+
+    const persons = [...this.state.persons];
+    persons[personIndex] = person;
+
+    this.setState({ persons: persons });
+  }
+
+  togglePersonsHandler = () => {
+    const doesShow = this.state.showPersons;
+    this.setState({showPersons: !doesShow});
+  }
+
+  deletePersonsHandler = (personIndex, id) => {
+    const persons = this.state.persons;
+    persons.splice(personIndex, 1);
+    this.setState({persons: persons});
+  }
+
+  render() {
+    let persons = null;
+    if (this.state.showPersons) {
+      persons = <Persons
+        persons={this.state.persons}
+        clicked={this.deletePersonsHandler}
+        changed={this.nameChangedHandler} />;
+    }
+
+    return (
+      <div className={classes.App}>
+        <Cockpit 
+          showPersons={this.state.showPersons} 
+          persons={this.state.persons}
+          clicked={this.togglePersonsHandler} />
+        {persons}
+      </div>
+    );
+  }
+}
+
+export default App;
